Add async_update_cache_of_contract_creator_on_demand for realtime fetcher as well
